heavily reduce usage of floating point numbers

+++ slam/include/slam/scan.h
@@ -17,6 +17,6 @@ void scan_reset(scan_t *scan);
 
 void scan_reset(scan_t *scan);
 
-void scan_add(scan_t *scan, unsigned short bearing, float range);
+void scan_add(scan_t *scan, uint16_t bearing, uint16_t range);
 
 #endif /* SLAM_SCAN_H_ */
